Move Freight Frenzy to teamcode.samples for reference
