+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/samples/CDDriveChassis.java
@@ -1,13 +1,11 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.samples;
 
 
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
-import com.qualcomm.robotcore.util.Range;
-
-import org.firstinspires.ftc.teamcode.*;
+// import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+// import com.qualcomm.robotcore.util.Range;
 
 
 /**
@@ -23,36 +21,39 @@ import org.firstinspires.ftc.teamcode.*;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-//TODO: Need to determine why this is a dependency in the project.
-
-public class CDDriveChassisOmni {
+public class CDDriveChassis {
 
     CDHardware robotHardware;
+    private ElapsedTime   runtime=new ElapsedTime();
 
-    public CDDriveChassisOmni (CDHardware theHardware) {
-
+    public CDDriveChassis(CDHardware theHardware) {
+        
         robotHardware = theHardware;
-
+        
         robotHardware.leftfrontmotor.setDirection(DcMotorSimple.Direction.FORWARD);
         robotHardware.leftrearmotor.setDirection(DcMotorSimple.Direction.FORWARD);
         robotHardware.rightfrontmotor.setDirection(DcMotorSimple.Direction.REVERSE);
         robotHardware.rightrearmotor.setDirection(DcMotorSimple.Direction.REVERSE);
-
+        
         robotHardware.leftfrontmotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         robotHardware.leftrearmotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         robotHardware.rightfrontmotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         robotHardware.rightrearmotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
     }
-    // drew is super cool and awesome and you will give him 20 dollars
-    public void setLeftPower (double pow) {
+
+    public void setLeftFrontPower (double pow) {
         robotHardware.leftfrontmotor.setPower(pow);
+    }
+    public void setLeftRearPower (double pow) {
         robotHardware.leftrearmotor.setPower(pow);
     }
-
-    public void setRightPower (double pow) {
+    public void setRightFrontPower (double pow) {
         robotHardware.rightfrontmotor.setPower(pow);
+    }
+    public void setRightRearPower (double pow) {
         robotHardware.rightrearmotor.setPower(pow);
     }
+
 }
