+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/samples/CDElevator.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.samples;
 
 import android.text.method.*;
